% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoordinateDescent}
\alias{CoordinateDescent}
\title{The algorithm of Coordinate Descent for LASSO}
\usage{
CoordinateDescent(X, y, lambda)
}
\arguments{
\item{X}{the observation vector}

\item{y}{the response vector}

\item{lambda}{the smoothing parameter}
}
\value{
the list which concludes the estimator beta, 
the corresponding values of objective functions and 
the numbers of iteration
}
\description{
The algorithm of Coordinate Descent for LASSO
}
\examples{
require(zhouyahomework1)
# Generate Data
n = 100
p = 500
sigma_noise = 0.5
beta = rep(0, p)
beta[1:6] = c(5,10,3,80,90,10)
XData = matrix(rnorm(n*p,sd=10), nrow = n, ncol=p)
X=XData
for (i in 1:500){
X[,i]=((X[,i]-sum(X[,i])/n))/((var(X[,i])*99/100)^0.5)
}
YData = X \%*\% beta + rnorm(n, sd = sigma_noise)
YData=YData-sum(YData)/100
XData=X
#Coordinate Descent for LASSO
betaHat1list = CoordinateDescent(XData,YData,1)
betaHat1=unlist(betaHat1list["betahat"])
}
