\name{zhouyahomework1-package}
\alias{zhouyahomework1-package}
\alias{zhouyahomework1}
\docType{package}
\title{
\packageTitle{zhouyahomework1}
}
\description{
\packageDescription{zhouyahomework1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{zhouyahomework1}
\packageIndices{zhouyahomework1}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{zhouyahomework1}

Maintainer: \packageMaintainer{zhouyahomework1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
